pyproject.toml
[tool.poetry]
name = "dbd_tooling"
version = "0.1.0"
description = ""
authors = ["Riadh"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
asyncio = "^3.4.3"
aiohttp = "^3.8.6"
minify-html = "^0.11.1"
pillow = "^10.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
